requirements.txt
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docopt==0.6.2
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
gymnasium==0.29.1
idna==3.8
imageio==2.35.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pettingzoo==1.24.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.5.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
protobuf==5.28.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pygame==2.6.0
Pygments==2.18.0
pymunk==6.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.35.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
SuperSuit==3.9.3
sympy==1.13.2
tensorboardX==2.6.2.2
tifffile==2024.8.30
tinycss2==1.3.0
tinyscaler==1.2.8
torch==2.4.1
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
